third_party/WebKit/LayoutTests/accessibility/canvas-fallback-content-expected.txt

Link Button    
1
2
Focusable
ARIA button
ARIA link


This test makes sure that focusable elements in canvas fallback content are accessible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

link1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXLink"


button1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"


text1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXTextField"


checkbox1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXCheckBox"


radio1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXRadioButton"


submit1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"


combobox1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXPopUpButton"


focusable1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXGenericContainer"


aria-button1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"


aria-link1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXLink"


link2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXLink"


button2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"


text2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXTextField"


checkbox2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXCheckBox"


radio2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXRadioButton"


submit2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"


combobox2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXPopUpButton"


focusable2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXGenericContainer"


aria-button2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"


aria-link2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXLink"


focusable1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"


focusable2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"


PASS successfullyParsed is true

TEST COMPLETE
